renames files camel case and corrects imports
