+++ public/js/views/modals/renderSettings.js
